requirements.txt
flask==3.1.1
openai==1.59.9
pandas==2.2.3
openpyxl==3.1.2
werkzeug==3.1.3 
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pdfplumber==0.11.4
python-multipart==0.0.12
python-dotenv==1.0.1 